Fix MTFP dense matmul accumulator overflow and wide-path red-team items
W1 (High): MTFP19 dense matmul accumulators widened from int64 to
__int128. The int64 path overflowed at K > 27 with MAX_VAL cells
(K=64 typical). With __int128, K_max ≈ 5e20 — effectively unlimited.
MTFP39 dense matmul K ≤ 41 constraint documented in header (no wider
hardware accumulator exists). Ternary matmul unaffected (add/sub only).

This bug was latent since the original glyph code (c021838) and carried
through the M4T extraction. Tests never caught it because all golden
values use small cells well below MAX_VAL.

W2: Added m4t_mtfp_w_vec_scale (scalar, delegates to m4t_mtfp_w_mul).
W3: Documented missing wide-path higher-level ops (bias_add, fan_in,
    layernorm) as intentionally deferred — wide path targets
    accumulation, not full transformer layers.
W4: Fixed header throughput claim from "half" to "~2.5× slower"
    (accounts for the 4-instruction compare+select clamp vs 2).
W5: Added test_w_vec_scale, test_w_vec_odd_n (n=3 NEON+tail, n=1
    pure scalar with saturation).

+++ m4t/src/m4t_mtfp.c
@@ -85,19 +85,22 @@ void m4t_mtfp_vec_scale(
 
 /* ── Dense MTFP × MTFP matmul ──────────────────────────────────────────── */
 
+/* Accumulator uses __int128 to avoid overflow at large K.
+ * With int64 accumulators, K > 27 with MAX_VAL cells would overflow.
+ * With __int128, K_max ≈ 5e20 — effectively unlimited. */
+
 static void matmul_row(
     m4t_mtfp_t* Y_row, const m4t_mtfp_t* X_row, const m4t_mtfp_t* W,
     int K, int N)
 {
     for (int j = 0; j < N; j++) {
-        int64_t acc = 0;
+        __int128 acc = 0;
         for (int k = 0; k < K; k++) {
-            acc += (int64_t)X_row[k] * (int64_t)W[(size_t)k * N + j];
+            acc += (__int128)X_row[k] * (__int128)W[(size_t)k * N + j];
         }
-        /* Round to nearest (ties away from zero), then rescale. */
         if (acc >= 0) acc += M4T_MTFP_SCALE / 2;
         else          acc -= M4T_MTFP_SCALE / 2;
-        Y_row[j] = m4t_mtfp_clamp64(acc / M4T_MTFP_SCALE);
+        Y_row[j] = m4t_mtfp_clamp64((int64_t)(acc / M4T_MTFP_SCALE));
     }
 }
 
@@ -107,13 +110,13 @@ static void matmul_bt_row(
 {
     for (int j = 0; j < N; j++) {
         const m4t_mtfp_t* wj = W + (size_t)j * K;
-        int64_t acc = 0;
+        __int128 acc = 0;
         for (int k = 0; k < K; k++) {
-            acc += (int64_t)X_row[k] * (int64_t)wj[k];
+            acc += (__int128)X_row[k] * (__int128)wj[k];
         }
         if (acc >= 0) acc += M4T_MTFP_SCALE / 2;
         else          acc -= M4T_MTFP_SCALE / 2;
-        Y_row[j] = m4t_mtfp_clamp64(acc / M4T_MTFP_SCALE);
+        Y_row[j] = m4t_mtfp_clamp64((int64_t)(acc / M4T_MTFP_SCALE));
     }
 }
